Update module resolution
I had made it so that app-time would stubbornly insist that it's
dependencies come from its own modules directory. This was the case in
dev using yarn link. However when installed in another project it needs
to be able to search through that projects node modules in order to
find things, since modules are most often installed in a flat tree
structure.